+++ CHANGELOG.md
@@ -38,6 +38,7 @@ x.x.x Release notes (yyyy-MM-dd)
 * Upgraded to Realm Core v5.13.0.
 * Upgraded to Realm Sync v3.14.14.
 * Stopped including headers from developers JDK when building the Android native module. ([#2223](https://github.com/realm/realm-js/pull/2223))
+* Added integration tests running in various environments (Node.js on MacOS and Linux, React Native on iOS and Android & Electron main and renderer processes on Linux). ([#2227](https://github.com/realm/realm-js/pull/2227)).
 
 2.22.0 Release notes (2019-1-10)
 =============================================================